Initialize column widths during creation instead of using a prop. listener. Fixed callbacks for "notify::width". Use g_object_set() instead of single functions to set column properties. Fixed and move cast for GUI_CELL_RENDERER_YPAD to definition. Minor style fixes.

+++ src/search_stats_gui2.c
@@ -308,7 +308,7 @@ static void add_column(
         GTK_CELL_RENDERER_TEXT(renderer), 1);
     g_object_set(renderer,
         "xalign", xalign,
-        "ypad", (gint) GUI_CELL_RENDERER_YPAD,
+        "ypad", GUI_CELL_RENDERER_YPAD,
         NULL);
 
     column = gtk_tree_view_column_new_with_attributes(
